Split out the outputting of the finishing of declarations as well
This is necessary because the generator is instantiated once for all the
capnp files requested. That means all state is carried on to the next
file, such as importId_. This caused a problem where imports were shared
among multiple files if capnp compile was called with them. This fix
makes 'capnp compile -ocara a.capnp b.capnp' equivalent to two separate
calls.